refactor(recipe): extract guarded raw YAML repair
